% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_daichou_idou.R
\name{read_idou_pref}
\alias{read_idou_pref}
\title{Read and tidy internal migration data for a specified prefecture in Japan.}
\usage{
read_idou_pref(path, pref_code, .tidy = FALSE)
}
\arguments{
\item{path}{Path to the Excel file containing the population migration data.}

\item{pref_code}{a character string specifying the code for the prefecture of
interest. This code should be a two-digit string (e.g. "01" for Hokkaido)}

\item{.tidy}{option. A logical flag indicating whether the output should
be in a tidy format or not.}
}
\description{
Read and tidy internal migration data for a specified prefecture in Japan.
}
\examples{
\dontrun{
read_idou_pref("b01036s.xlsx", pref_code = "36")
}
}
